AC-300 implemented Ed's suggestions
